dependency object working
